refactor: Add check for signed-in user in question page
This commit adds a check for a signed-in user in the question page. Previously, the page assumed that the user must be signed in to view questions. However, we also want to allow users to view questions without signing in. This change ensures that the `mongoUser` variable is only assigned if `clerkId` is provided, indicating that the user is signed in.